runs/codex.kit/REQ-001/requirements.txt
fastapi==0.111.0
httpx==0.27.2
pydantic==2.8.2
python-jose[cryptography]==3.3.0
uvicorn==0.30.1
pytest==8.3.2
pytest-asyncio==0.23.7
anyio==4.4.0